Refactor the structure of agent tools
